+++ pkg/client/history/table_test.go
@@ -1,3 +1,4 @@
+//go:build history
 // +build history
 
 package history_test