+++ pkg/policy/filter.go
@@ -6,7 +6,10 @@ import (
 	"github.com/cloudquery/cloudquery/pkg/config"
 )
 
-func FilterPolicies(args []string, configPolicies []*config.Policy, policyName, subPath string) ([]*config.Policy, error) {
+// TODO: move PolicyName to be also as part of args i.e first arg is policy name second arg is subpath, the rest is taken from config,
+// TODO: if first arg isn't found in config, execute remote hub download only policy
+
+func FilterPolicies(args []string, configPolicies []*config.Policy, policyName string) ([]*config.Policy, error) {
 	var policies []*config.Policy
 
 	if len(args) > 0 {
@@ -15,7 +18,9 @@ func FilterPolicies(args []string, configPolicies []*config.Policy, policyName, 
 			return nil, err
 		}
 		policyConfig, err := remotePolicy.ToPolicyConfig()
-		policyConfig.SubPath = subPath
+		if len(args) == 2 {
+			policyConfig.SubPath = args[1]
+		}
 		if err != nil {
 			return nil, err
 		}
@@ -27,8 +32,7 @@ func FilterPolicies(args []string, configPolicies []*config.Policy, policyName, 
 	if len(policies) == 0 {
 		return nil, fmt.Errorf(`
 Could not find policies to run.
-Please add policy to block to your config file.
-`)
+Please add policy to block to your config file`)
 	}
 	policiesToRun := make([]*config.Policy, 0)
 
@@ -37,10 +41,6 @@ Please add policy to block to your config file.
 		if policyName != "" {
 			// request to run only specific policy
 			if policyName == p.Name {
-				// override subPath if specified
-				if subPath != "" {
-					p.SubPath = subPath
-				}
 				policiesToRun = append(policiesToRun, p)
 				break
 			}