+++ pkg/ui/console/client.go
@@ -6,6 +6,7 @@ import (
 	"errors"
 	"fmt"
 	"os"
+	"path"
 	"strings"
 	"time"
 
@@ -150,10 +151,9 @@ func (c Client) DownloadPolicy(ctx context.Context, args []string) error {
 	return nil
 }
 
-func (c Client) RunPolicies(ctx context.Context, args []string, policyName, outputDir, subPath string, stopOnFailure, skipVersioning, failOnViolation, noResults bool) error {
+func (c Client) RunPolicies(ctx context.Context, args []string, policyName, outputDir string, stopOnFailure, skipVersioning, failOnViolation, noResults bool) error {
 	c.c.Logger.Debug("Received params:", "args", args, "policyName", policyName, "outputDir", outputDir, "stopOnFailure", stopOnFailure, "skipVersioning", skipVersioning, "failOnViolation", failOnViolation, "noResults", noResults)
-
-	policiesToRun, err := policy.FilterPolicies(args, c.cfg.Policies, policyName, subPath)
+	policiesToRun, err := policy.FilterPolicies(args, c.cfg.Policies, policyName)
 
 	if err != nil {
 		ui.ColorizedOutput(ui.ColorError, err.Error())
@@ -201,6 +201,48 @@ func (c Client) RunPolicies(ctx context.Context, args []string, policyName, outp
 
 	ui.ColorizedOutput(ui.ColorProgress, "Finished policies run...\n\n")
 	return nil
+}
+
+func (c Client) DescribePolicies(ctx context.Context, args []string, policyName string, skipVersioning bool) error {
+	policiesToDescribe, err := policy.FilterPolicies(args, c.cfg.Policies, policyName)
+	if err != nil {
+		ui.ColorizedOutput(ui.ColorError, err.Error())
+		return err
+	}
+	req := &client.PoliciesRunRequest{
+		Policies:       policiesToDescribe,
+		PolicyName:     policyName,
+		SkipVersioning: skipVersioning,
+	}
+	return c.describePolicies(ctx, req)
+}
+
+func (c Client) describePolicies(ctx context.Context, req *client.PoliciesRunRequest) error {
+	policies, err := c.c.LoadPolicies(ctx, req)
+	if err != nil {
+		ui.ColorizedOutput(ui.ColorError, err.Error())
+		return fmt.Errorf("failed to load policies: %w", err)
+	}
+	ui.ColorizedOutput(ui.ColorHeader, "Describe Policy %s output:\n\n", req.PolicyName)
+	t := &Table{writer: tablewriter.NewWriter(os.Stdout)}
+	t.SetHeaders("Path", "Description")
+	buildDescribePolicyTable(t, policies, "")
+	t.Render()
+	ui.ColorizedOutput(ui.ColorInfo, "To execute any policy use the path defined in the table above.\nFor example `cloudquery policy run %s %s`", req.PolicyName, getNestedPolicyExample(policies[0], ""))
+	return nil
+}
+
+func getNestedPolicyExample(p *policy.Policy, policyPath string) string {
+	if len(p.Policies) > 0 {
+		return getNestedPolicyExample(p.Policies[0], path.Join(policyPath, strings.ToLower(p.Name)))
+	}
+	return policyPath
+}
+func buildDescribePolicyTable(t ui.Table, pp policy.Policies, policyPath string) {
+	for _, p := range pp {
+		t.Append(path.Join(policyPath, strings.ToLower(p.Name)), p.Description)
+		buildDescribePolicyTable(t, p.Policies, path.Join(policyPath, strings.ToLower(p.Name)))
+	}
 }
 
 func (c Client) CallModule(ctx context.Context, req ModuleCallRequest) error {